CCO-xAPP/services/trainer/requirements.txt
kafka-python
pandas
scikit-learn
xgboost
joblib
numpy
matplotlib
seaborn
joblib